src/main/java/org/example/model/Triangle.java
package org.example.model;

import jakarta.persistence.Entity;
import jakarta.persistence.Table;

@Entity
@Table(name = "Triangle")
public class Triangle extends Shape {
    private int a;
    private int b;
    private int c;

    public Triangle() {
    }

    public Triangle(int a, int b, int c, Color color) {
        super(color);
        this.a = a;
        this.b = b;
        this.c = c;
        this.perimeter = a + b + c;
        double halfPerimeter = 0.5 * this.perimeter;
        this.area = Math.sqrt((halfPerimeter * (halfPerimeter - this.a) * (halfPerimeter - this.b) * (halfPerimeter - this.c)));
    }

    @Override
    public double getArea() {
        return this.area;
    }

    @Override
    public double getPerimeter() {
        return this.perimeter;
    }

    public int getA() {
        return a;
    }

    public int getB() {
        return b;
    }

    public int getC() {
        return c;
    }
}
